wildberries_robot/requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==37.0.2
cssselect==1.1.0
Deprecated==1.2.13
Django==3.2.13
django-annoying==0.10.6
django-celery-beat==2.2.1
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==3.0.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
fake-useragent==0.1.11
flower==1.0.0
gunicorn==20.1.0
h11==0.13.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
parse==1.19.0
pep517==0.12.0
phonenumberslite==8.12.48
Pillow==9.1.0
pip-compile-multi==2.4.5
pip-tools==6.6.0
pkg_resources==0.0.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-html==0.10.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
tornado==6.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.5.4
websockets==10.3
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
